+++ src/app/shared/cohort/cohort.service.spec.ts
@@ -13,9 +13,10 @@ import {
   HttpTestingController,
   HttpClientTestingModule
 } from '@angular/common/http/testing';
-import { Criteria } from '../../types/search-parameters';
+import { Criteria, ResourceTypeCriteria } from '../../types/search-parameters';
 import { configureTestingModule } from 'src/test/helpers';
 import { RouterTestingModule } from '@angular/router/testing';
+import { of } from 'rxjs';
 
 describe('CohortService', () => {
   let cohort: CohortService;
@@ -264,10 +265,358 @@ describe('CohortService', () => {
     );
     cohort.updateOldFormatCriteria(criteria);
     expect(criteria.rules[0].rules[1].field.value).toEqual(
-      jasmine.objectContaining({ observationDataType: 'CodeableConcept' })
+      jasmine.objectContaining({observationDataType: 'CodeableConcept'})
     );
     expect(Object.keys(criteria.rules[0].rules[1].field)).not.toContain(
       'observationDataType'
     );
   });
+
+
+  /**
+   * Creates an example of criteria for a resource type.
+   * @param resourceType resource type with a prefix or suffix describing its modification
+   * @return a new object describes criteria to select Patient resources.
+   */
+  function getCriteriaFor(
+    resourceType: 'Patient' | 'ResearchStudy' | 'otherResearchStudy'
+      | 'Condition' | 'PatientForResearchStudy' | 'PatientForOtherResearchStudy'
+      | 'PatientCombinedWithPatientForResearchStudy' | 'nonexistentResearchStudy'
+      | 'PatientForCondition'
+  ): ResourceTypeCriteria {
+    switch (resourceType) {
+      case 'Patient':
+        return {
+          condition: 'and',
+          resourceType: 'Patient',
+          rules: [
+            {
+              field: {
+                element: 'gender',
+                value: {
+                  codes: ['female'],
+                  items: ['Female']
+                }
+              }
+            }
+          ]
+        };
+      case 'ResearchStudy':
+      case 'otherResearchStudy':
+      case 'nonexistentResearchStudy':
+        return {
+          condition: 'and',
+          resourceType: 'ResearchStudy',
+          rules: [
+            {
+              field: {
+                element: 'title',
+                value: {
+                  testValuePrefix: '',
+                  testValueModifier: '',
+                  testValue: {
+                    ResearchStudy: 'a',
+                    otherResearchStudy: 'b',
+                    nonexistentResearchStudy: 'nonexistent'
+                  }[resourceType],
+                  testValueUnit: '',
+                  testValuePrefix2: '',
+                  testValue2: '',
+                  observationDataType: 'String'
+                }
+              }
+            }
+          ]
+        };
+      case 'PatientForResearchStudy':
+      case 'PatientForOtherResearchStudy':
+        return {
+          condition: 'and',
+          resourceType: 'Patient',
+          rules: [
+            {
+              field: {
+                element: '_has:ResearchSubject:individual:study',
+                value: resourceType === 'PatientForResearchStudy' ? 'study-1,study-2' : 'study-3'
+              }
+            }
+          ]
+        };
+      case 'PatientCombinedWithPatientForResearchStudy':
+        return {
+          condition: 'and',
+          resourceType: 'Patient',
+          rules: [
+            {
+              field: {
+                element: 'gender',
+                value: {
+                  codes: ['female'],
+                  items: ['Female']
+                }
+              }
+            },
+            {
+              field: {
+                element: '_has:ResearchSubject:individual:study',
+                value: 'study-1,study-2'
+              }
+            }
+          ]
+        };
+      case 'Condition':
+        return {
+          condition: 'and',
+          resourceType: 'Condition',
+          rules: [
+            {
+              field: {
+                element: 'code',
+                value: {
+                  codes: ['276.3'],
+                  items: ['ALKALOSIS']
+                }
+              }
+            }
+          ]
+        };
+      case 'PatientForCondition':
+        return {
+          condition: 'and',
+          resourceType: 'Patient',
+          rules: [
+            {
+              field: {
+                element: '_has:Condition:subject:code',
+                value: '276.3'
+              }
+            }
+          ]
+        };
+    }
+  }
+
+
+  describe('simplifyCriteriaTree', () => {
+    it('should move ANDed child criteria to the parent criteria', () => {
+      expect(
+        cohort.simplifyCriteriaTree({
+          condition: 'and',
+          rules: [
+            getCriteriaFor('Patient'),
+            {
+              condition: 'and',
+              rules: [
+                getCriteriaFor('ResearchStudy'),
+                getCriteriaFor('Condition')
+              ]
+            }
+          ]
+        })
+      ).toEqual({
+        condition: 'and',
+        rules: [
+          getCriteriaFor('Patient'),
+          getCriteriaFor('ResearchStudy'),
+          getCriteriaFor('Condition')
+        ]
+      });
+    });
+
+
+    it('should not move ORed child criteria to the parent criteria', () => {
+      expect(
+        cohort.simplifyCriteriaTree({
+          condition: 'and',
+          rules: [
+            getCriteriaFor('Patient'),
+            {
+              condition: 'or',
+              rules: [
+                getCriteriaFor('ResearchStudy'),
+                getCriteriaFor('Condition')
+              ]
+            }
+          ]
+        })
+      ).toEqual({
+        condition: 'and',
+        rules: [
+          getCriteriaFor('Patient'),
+          {
+            condition: 'or',
+            rules: [
+              getCriteriaFor('ResearchStudy'),
+              getCriteriaFor('Condition')
+            ]
+          }
+        ]
+      });
+
+      expect(
+        cohort.simplifyCriteriaTree({
+          condition: 'or',
+          rules: [
+            getCriteriaFor('Patient'),
+            {
+              condition: 'and',
+              rules: [
+                getCriteriaFor('ResearchStudy'),
+                getCriteriaFor('Condition')
+              ]
+            }
+          ]
+        })
+      ).toEqual({
+        condition: 'or',
+        rules: [
+          getCriteriaFor('Patient'),
+          {
+            condition: 'and',
+            rules: [
+              getCriteriaFor('ResearchStudy'),
+              getCriteriaFor('Condition')
+            ]
+          }
+        ]
+      });
+    });
+  });
+
+
+  describe('replaceOtherResourceCriteriaWithPatientCriteria', () => {
+    it('should replace ResearchStudy criteria with patient criteria', (done) => {
+      cohort.replaceOtherResourceCriteriaWithPatientCriteria(of({
+        'condition': 'and',
+        'rules': [
+          {
+            'condition': 'and',
+            'rules': [
+              getCriteriaFor('Patient'),
+              getCriteriaFor('ResearchStudy'),
+              getCriteriaFor('Condition')
+            ]
+          },
+          {
+            'condition': 'or',
+            'rules': [
+              getCriteriaFor('Patient'),
+              getCriteriaFor('otherResearchStudy'),
+              {
+                condition: 'and',
+                rules: [
+                  getCriteriaFor('Condition'),
+                ]
+              },
+              {
+                condition: 'and',
+                rules: [
+                  getCriteriaFor('Condition'),
+                  getCriteriaFor('nonexistentResearchStudy'),
+                ]
+              }
+            ]
+          }
+        ]
+      })).subscribe((criteria) => {
+        expect(criteria).toEqual({
+          'condition': 'and',
+          'rules': [
+            {
+              'condition': 'and',
+              'rules': [
+                getCriteriaFor('Patient'),
+                getCriteriaFor('PatientForResearchStudy'),
+                getCriteriaFor('PatientForCondition')
+              ]
+            },
+            {
+              'condition': 'or',
+              'rules': [
+                getCriteriaFor('Patient'),
+                getCriteriaFor('PatientForOtherResearchStudy'),
+                {
+                  condition: 'and',
+                  rules: [
+                    getCriteriaFor('PatientForCondition'),
+                  ]
+                }
+              ]
+            }
+          ]
+        });
+
+        done();
+      });
+
+      mockHttp
+        .expectOne(
+          '$fhir/ResearchStudy?_elements=id&_has:ResearchSubject:study:status=candidate,eligible,follow-up,ineligible,not-registered,off-study,on-study,on-study-intervention,on-study-observation,pending-on-study,potential-candidate,screening,withdrawn&title=a'
+        )
+        .flush({
+          entry: [{resource: {id: 'study-1'}}, {resource: {id: 'study-2'}}]
+        });
+      mockHttp
+        .expectOne(
+          '$fhir/ResearchStudy?_elements=id&_has:ResearchSubject:study:status=candidate,eligible,follow-up,ineligible,not-registered,off-study,on-study,on-study-intervention,on-study-observation,pending-on-study,potential-candidate,screening,withdrawn&title=b'
+        )
+        .flush({
+          entry: [{resource: {id: 'study-3'}}]
+        });
+      mockHttp
+        .expectOne(
+          '$fhir/ResearchStudy?_elements=id&_has:ResearchSubject:study:status=candidate,eligible,follow-up,ineligible,not-registered,off-study,on-study,on-study-intervention,on-study-observation,pending-on-study,potential-candidate,screening,withdrawn&title=nonexistent'
+        )
+        .flush({
+          entry: []
+        });
+    });
+  });
+
+
+  describe('combineANDedCriteriaForPatient', () => {
+    it('should combine ANDed criteria for Patient', () => {
+      expect(
+        cohort.combineANDedCriteriaForPatient({
+          condition: 'and',
+          rules: [
+            {
+              condition: 'and',
+              rules: [
+                getCriteriaFor('Patient'),
+                getCriteriaFor('PatientForResearchStudy'),
+                getCriteriaFor('Condition')]
+            },
+            {
+              condition: 'or',
+              rules: [
+                getCriteriaFor('Patient'),
+                getCriteriaFor('PatientForOtherResearchStudy'),
+                getCriteriaFor('Condition')
+              ]
+            }
+          ]
+        })
+      ).toEqual({
+        condition: 'and',
+        rules: [
+          {
+            condition: 'and',
+            rules: [
+              getCriteriaFor('PatientCombinedWithPatientForResearchStudy'),
+              getCriteriaFor('Condition')]
+          },
+          {
+            condition: 'or',
+            rules: [
+              getCriteriaFor('Patient'),
+              getCriteriaFor('PatientForOtherResearchStudy'),
+              getCriteriaFor('Condition')
+            ]
+          }
+        ]
+      });
+    });
+  });
+
 });